% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar chart from a computed tables}
\usage{
plot_bar(
  x,
  chart_colors,
  legend_location,
  include_bar_text = FALSE,
  percent = FALSE
)
}
\arguments{
\item{x}{A table with a group column 'text' and a count column 'numeric' or 'integer' in that order.}

\item{chart_colors}{A vector of colors to use with the category chart.}

\item{legend_location}{Can be one of "none", "bottom", "right", "left". Defaults to none.}

\item{include_bar_text}{Whether to include the category text within the bar. Defaults to TRUE}

\item{percent}{Whether to the categories display counts or percent.}
}
\description{
Bar chart from a computed tables
}
