% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category.R
\name{plot_category}
\alias{plot_category}
\title{Category charts from computed tables}
\usage{
plot_category(
  x,
  plot_type = "pie",
  legend_location = "none",
  chart_colors = NULL,
  include_pie_numbers = FALSE,
  include_bar_text = TRUE,
  waffle_cols = 12,
  percent = FALSE
)
}
\arguments{
\item{x}{A table with a group column 'text' and a count column 'numeric' or 'integer' in that order.}

\item{plot_type}{One of four optins. Pie Chart 'pie', Bar Chart 'bar', Treemap 'tree', Waffle 'waffle'}

\item{legend_location}{Can be one of "none", "bottom", "right", "left". Defaults to none.}

\item{chart_colors}{A vector of colors to use with the category chart.  Defaults to NULL. If NULL a six color pallete is used.}

\item{include_pie_numbers}{Whether to include the numbers within the slices of the pie chart. Defaults to FALSE.}

\item{include_bar_text}{Whether to include the category text within the bar. Defaults to TRUE}

\item{waffle_cols}{The number of columns of the waffle chart.}

\item{percent}{Whether to the categories display counts or percent.}
}
\description{
Category charts from computed tables
}
