% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category.R
\name{plot_pie}
\alias{plot_pie}
\title{Pie charts from computed tables}
\usage{
plot_pie(
  x,
  chart_colors,
  legend_location,
  include_pie_numbers = FALSE,
  percent = FALSE
)
}
\arguments{
\item{x}{A table with a group column 'text' and a count column 'numeric' or 'integer' in that order.}

\item{chart_colors}{A vector of colors to use with the category chart.}

\item{legend_location}{Can be one of "none", "bottom", "right", "left". Defaults to none.}

\item{include_pie_numbers}{Whether to include the numbers within the slices of the pie chart. Defaults to FALSE.}

\item{percent}{Whether to the categories display counts or percent.}
}
\description{
Pie charts from computed tables
}
