% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_category.R
\name{plot_tree}
\alias{plot_tree}
\title{Treemap chart from a computed tables}
\usage{
plot_tree(x, chart_colors, legend_location, percent = FALSE)
}
\arguments{
\item{x}{A table with a group column 'text' and a count column 'numeric' or 'integer' in that order.}

\item{chart_colors}{A vector of colors to use with the category chart.  Defaults to NULL. If NULL a six color pallete is used.}

\item{legend_location}{Can be one of "none", "bottom", "right", "left". Defaults to none.}

\item{percent}{Whether to the categories display counts or percent.}
}
\description{
Treemap chart from a computed tables
}
